requirements.txt
# Core dependencies
boto3>=1.26.0
botocore>=1.29.0
requests>=2.28.0
pyyaml>=6.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0

# Code quality and validation
flake8>=5.0.0
black>=22.0.0
bandit>=1.7.0
cfn-lint>=0.70.0

# Development dependencies
python-dotenv>=0.20.0
